Put the functions in a different file for better SoC
