+++ src/app.js
@@ -7,7 +7,7 @@ import { getMovieRating } from './rating.js';
 import cmsAdapterRating from './cmsAdapterRating.js';
 import cmsAdapter from './cmsAdapterTop5Movies.js';
 import cmsReviews from '../static/cmsReviews.js';
-import jsonwebtoken from 'jsonwebtoken';
+import { reviewAdapter } from './sendReviews.js';
 
 export default function initialize(api) {
   const app = express();
@@ -104,7 +104,6 @@ export default function initialize(api) {
       const movieID = req.params.id;
       const screenings = await cmsScreening.loadScreeningsID(req, res);
       const rating = await getMovieRating(cmsAdapterRating, movieID);
-      console.log(rating);
 
       if (!screenings) {
         throw new Error('Array does not contain any screenings.');
@@ -120,93 +119,12 @@ export default function initialize(api) {
     }
   });
 
-  const LOGIN = [
-    {
-      username: 'viktor',
-      password: 'wilma',
-    },
-    {
-      username: 'rikard',
-      password: 'rikard3209',
-    },
-    {
-      username: 'johnny',
-      password: 'zidust',
-    },
-    {
-      username: 'johannes',
-      password: 'spinalglitter38',
-    },
-    {
-      username: 'louise',
-      password: 'lillamylouise',
-    },
-    {
-      username: 'emily',
-      password: 'emily_92251',
-    },
-    {
-      username: 'oscar',
-      password: 'oscar0709',
-    },
-  ];
-  const USERNAME = "admin";
-  const PASSWORD = "secret";
-  const SECRET = "averylongpasswordthattheonlyonewhoknowswhatitisisthiscomputercodeblock";
-
   app.post("/api/login", (req, res) => {
-    const authHeader = req.headers.authorization;
-    const b64credentials = authHeader.slice(6);
-    const credentials = atob(b64credentials);
-    const fields = credentials.split(":");
-    const username = fields[0];
-    const password = fields[1];
-    console.log("Username:", username);
-    console.log("Password:", password);
-
-    if (
-      (
-        username == LOGIN[0].username && password == LOGIN[0].password ||
-        username == LOGIN[1].username && password == LOGIN[1].password ||
-        username == LOGIN[2].username && password == LOGIN[2].password ||
-        username == LOGIN[3].username && password == LOGIN[3].password ||
-        username == LOGIN[4].username && password == LOGIN[4].password ||
-        username == LOGIN[5].username && password == LOGIN[5].password ||
-        username == LOGIN[6].username && password == LOGIN[6].password 
-      )
-    ) {
-      const jwt = jsonwebtoken.sign(
-        {
-          username: username,
-          role: 'reviewer',
-        },
-        SECRET
-      );
-
-      res.status(200).json({
-        ok: true,
-        token: jwt,
-      });
-    } else {
-      res.status(401).end();
-    }
+    reviewAdapter.loadLogin(req, res);
   });
 
-  app.get("/api/reviews", (req, res) => {
-    const authHeader = req.headers.authorization;
-    const token = authHeader?.slice(7);
-
-    try {
-      const payload = jsonwebtoken.verify(token, SECRET);
-      res.status(200).json({
-        ok: true,
-      });
-    } catch (error) {
-      res.status(401).json({
-        ok: false,
-        error: "not allowed",
-      });
-    }
+  app.get('/api/reviews', (req, res) => {
+    reviewAdapter.sendReviewAccess(req, res);
   });
 
   app.use('/static', express.static('./static'));
